Rename StandardLinker.Config to StandardConfig
We will move it into a different package, so we need to separate it out.
